backend-api/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
amqp==5.1.1
    # via kombu
asgiref==3.4.1
    # via django
assertpy==1.1
    # via behave-restful
attrs==21.2.0
    # via jsonschema
beautifulsoup4==4.10.0
    # via behave-django
behave==1.2.6
    # via
    #   behave-django
    #   behave-restful
behave-django==1.4.0
    # via -r requirements.in
behave-restful==0.4.1
    # via -r requirements.in
billiard==3.6.4.0
    # via celery
boto3==1.24.7
    # via -r requirements.in
botocore==1.27.7
    # via
    #   boto3
    #   s3transfer
cachecontrol==0.12.11
    # via firebase-admin
cachetools==5.0.0
    # via google-auth
celery==5.2.6
    # via -r requirements.in
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via cryptography
charset-normalizer==2.0.9
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
coreapi==2.3.3
    # via djangorestframework-stubs
coreschema==0.0.4
    # via coreapi
cryptography==36.0.2
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
dj-rest-auth==2.2.3
    # via -r requirements.in
django==3.2
    # via
    #   -r requirements.in
    #   dj-rest-auth
    #   django-allauth
    #   django-cors-headers
    #   django-filter
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework
    #   djangorestframework-jsonapi
    #   djangorestframework-simplejwt
    #   fcm-django
django-allauth==0.50.0
    # via -r requirements.in
django-cors-headers==3.13.0
    # via -r requirements.in
django-filter==21.1
    # via -r requirements.in
django-stubs==1.9.0
    # via
    #   -r requirements.in
    #   djangorestframework-stubs
django-stubs-ext==0.3.1
    # via django-stubs
djangorestframework==3.12.4
    # via
    #   -r requirements.in
    #   dj-rest-auth
    #   djangorestframework-jsonapi
    #   djangorestframework-simplejwt
djangorestframework-jsonapi==4.3.0
    # via -r requirements.in
djangorestframework-simplejwt==4.8.0
    # via -r requirements.in
djangorestframework-stubs==1.4.0
    # via -r requirements.in
fcm-django==1.0.11
    # via -r requirements.in
firebase-admin==5.2.0
    # via fcm-django
google-api-core[grpc]==2.7.2
    # via
    #   firebase-admin
    #   google-api-python-client
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-storage
google-api-python-client==2.46.0
    # via firebase-admin
google-auth==2.6.6
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-core
    #   google-cloud-storage
google-auth-httplib2==0.1.0
    # via google-api-python-client
google-cloud-core==2.3.0
    # via
    #   google-cloud-firestore
    #   google-cloud-storage
google-cloud-firestore==2.4.0
    # via firebase-admin
google-cloud-storage==2.3.0
    # via firebase-admin
google-crc32c==1.3.0
    # via google-resumable-media
google-resumable-media==2.3.2
    # via google-cloud-storage
googleapis-common-protos==1.56.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.44.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.44.0
    # via google-api-core
httplib2==0.20.4
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.3
    # via requests
importlib-resources==5.4.0
    # via jsonschema
inflection==0.5.1
    # via djangorestframework-jsonapi
itypes==1.2.0
    # via coreapi
jinja2==3.0.3
    # via coreschema
jmespath==1.0.0
    # via
    #   boto3
    #   botocore
jsonpath==0.82
    # via behave-restful
jsonref==0.2
    # via behave-restful
jsonschema==4.3.2
    # via behave-restful
kombu==5.2.4
    # via celery
markupsafe==2.0.1
    # via jinja2
msgpack==1.0.3
    # via cachecontrol
mypy==0.920
    # via
    #   -r requirements.in
    #   django-stubs
    #   djangorestframework-stubs
mypy-extensions==0.4.3
    # via mypy
oauthlib==3.2.0
    # via requests-oauthlib
parse==1.19.0
    # via
    #   behave
    #   parse-type
parse-type==0.5.2
    # via behave
prompt-toolkit==3.0.29
    # via click-repl
proto-plus==1.20.3
    # via google-cloud-firestore
protobuf==3.20.1
    # via
    #   google-api-core
    #   google-cloud-storage
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
psycopg2-binary==2.8.5
    # via -r requirements.in
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.3.0
    # via
    #   django-allauth
    #   djangorestframework-simplejwt
pyparsing==3.0.8
    # via httplib2
pyrsistent==0.18.0
    # via jsonschema
python-dateutil==2.8.2
    # via botocore
python3-openid==3.2.0
    # via django-allauth
pytz==2021.3
    # via
    #   celery
    #   django
pyyaml==6.0
    # via behave-restful
requests==2.26.0
    # via
    #   behave-restful
    #   cachecontrol
    #   coreapi
    #   django-allauth
    #   djangorestframework-stubs
    #   google-api-core
    #   google-cloud-storage
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via django-allauth
rsa==4.8
    # via google-auth
s3transfer==0.6.0
    # via boto3
six==1.16.0
    # via
    #   behave
    #   click-repl
    #   google-auth
    #   google-auth-httplib2
    #   grpcio
    #   parse-type
    #   python-dateutil
soupsieve==2.3.1
    # via beautifulsoup4
sqlparse==0.4.2
    # via django
toml==0.10.2
    # via django-stubs
tomli==2.0.0
    # via mypy
types-pytz==2021.3.3
    # via django-stubs
types-pyyaml==6.0.1
    # via django-stubs
typing-extensions==4.0.1
    # via
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework-stubs
    #   mypy
uritemplate==4.1.1
    # via
    #   coreapi
    #   google-api-python-client
urllib3==1.26.7
    # via
    #   botocore
    #   requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
zipp==3.6.0
    # via importlib-resources
